+++ source/Internal/Framework/Module/Configuration/DataObject/ModuleConfiguration.php
@@ -241,7 +241,6 @@ class ModuleConfiguration
     }
 
     /**
-     * @deprecated will be completely removed in v7.0
      * @return bool
      */
     public function isActivated(): bool