+++ source/Internal/Framework/Module/Setup/Service/ModuleActivationServiceInterface.php
@@ -9,23 +9,17 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Service;
 
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Exception\ModuleSetupException;
-
 interface ModuleActivationServiceInterface
 {
     /**
      * @param string $moduleId
      * @param int    $shopId
-     *
-     * @throws ModuleSetupException
      */
     public function activate(string $moduleId, int $shopId);
 
     /**
      * @param string $moduleId
      * @param int    $shopId
-     *
-     * @throws ModuleSetupException
      */
     public function deactivate(string $moduleId, int $shopId);
 }